updated code.

+++ src/main/java/com/softvision/ipm/pms/role/repo/RoleRepository.java
@@ -7,13 +7,13 @@ import com.softvision.ipm.pms.common.repo.AbstractRepository;
 @Repository
 public class RoleRepository extends AbstractRepository {
 
-	public int assign(Long employeeId, int roleId) {
+	public int assign(Integer employeeId, int roleId) {
 		return jdbcTemplate.update(
 				"INSERT INTO employee_role(id, role_id, employee_id) VALUES (nextval('emp_role_id_seq'), ?,?)",
 				 roleId,employeeId);
 	}
 
-	public int remove(Long employeeId, int roleId) {
+	public int remove(Integer employeeId, int roleId) {
 		return jdbcTemplate.update("delete from employee_role where employee_id=? and role_id=?", employeeId, roleId);
 	}
