+++ src/main/java/com/softvision/ipm/pms/role/rest/RoleRest.java
@@ -6,6 +6,7 @@ import javax.validation.constraints.NotNull;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
@@ -46,10 +47,11 @@ public class RoleRest {
 	public @ResponseBody List<Role> getRolesBylogin(@PathVariable(value = "loginId", required = true) String loginId) {
 		return roleService.getRolesbyLoginId(loginId);
 	}
+
 	@PreAuthorize ( "hasRole('Admin')" )
 	@RequestMapping(value = "/delete/{employeeId}/{roleId}", method = RequestMethod.PUT)
-	public Result delete(@PathVariable(required = true) @NotNull long employeeId,
-			@PathVariable(required = true) @NotNull int roleId) {
+	public Result delete(@PathVariable(required = true) @NotNull Integer employeeId,
+			@PathVariable(required = true) @NotNull Integer roleId) {
 		Result result = new Result();
 		try {
 			roleService.removeRole(employeeId, roleId);
@@ -63,8 +65,8 @@ public class RoleRest {
 	}
 	@PreAuthorize ( "hasRole('Admin')" )
 	@RequestMapping(value = "/assign/{employeeId}/{roleId}", method = RequestMethod.PUT)
-	public Result assign(@PathVariable(required = true) @NotNull long employeeId,
-			@PathVariable(required = true) @NotNull int roleId) {
+	public Result assign(@PathVariable(required = true) @NotNull Integer employeeId,
+			@PathVariable(required = true) @NotNull Integer roleId) {
 		Result result = new Result();
 		try {
 			roleService.assignRole(employeeId, roleId);
